+++ shell/edit/provisioning.cattle.io.cluster/tabs/etcd/index.vue
@@ -3,7 +3,7 @@ import { LabeledInput } from '@components/Form/LabeledInput';
 
 import { RadioGroup } from '@components/Form/Radio';
 
-import S3Config from '@shell/edit/provisioning.cattle.io.cluster/S3Config';
+import S3Config from '@shell/edit/provisioning.cattle.io.cluster/tabs/etcd/S3Config';
 import UnitInput from '@shell/components/form/UnitInput';
 
 export default {
