+++ shell/edit/provisioning.cattle.io.cluster/tabs/registries/RegistryConfigs.vue
