+++ shell/edit/provisioning.cattle.io.cluster/tabs/registries/RegistryMirrors.vue
