+++ shell/edit/provisioning.cattle.io.cluster/tabs/upgrade/DrainOptions.vue
